Add files via upload
The fog shader file is located in the /shaders directory, with a test scene inclusive. The test scene shows a background of the game named Alto's Odyssey(as it offers quite beautiful backgrounds :P), with the shader overlay on it. You can change the parameters and color of the fog shader to match your background.